renamed StatefulEventBus to EventBusInteractionContext
