+++ myeslib2-jdbi/src/main/java/org/myeslib/jdbi/function/eventbus/EventBusInteractionContext.java
@@ -1,5 +1,6 @@
-package org.myeslib.jdbi.function;
+package org.myeslib.jdbi.function.eventbus;
 
+import com.google.common.eventbus.EventBus;
 import org.myeslib.core.AggregateRoot;
 import org.myeslib.core.Command;
 import org.myeslib.core.Event;
@@ -13,41 +14,46 @@ import java.util.List;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-public class MultiMethodInteractionContext implements InteractionContext {
+public class EventBusInteractionContext implements InteractionContext {
 
-    private static final Logger logger = LoggerFactory.getLogger(MultiMethodInteractionContext.class);
+    private static final Logger logger = LoggerFactory.getLogger(EventBusInteractionContext.class);
     private final AggregateRoot aggregateRoot;
-    private final MultiMethod mm ;
+    private final EventBus bus;
     private final List<Event> events;
     private final List<Command> commands;
 
-    public MultiMethodInteractionContext(AggregateRoot aggregateRoot) {
-        this.mm = MultiMethod.getMultiMethod(aggregateRoot.getClass(), "on");
+    public EventBusInteractionContext(AggregateRoot aggregateRoot) {
+        this.bus = new EventBus();
         this.events = new ArrayList<>();
         this.commands = new ArrayList<>();
         checkNotNull(aggregateRoot);
         this.aggregateRoot = aggregateRoot;
+        bus.register(aggregateRoot);
     }
 
     @Override
     public void apply(Event event) {
         checkNotNull(event);
         logger.debug("applying event {} generated by an operation on {}", event, aggregateRoot);
-        _applyEventsOn(aggregateRoot, event);
+        bus.post(event);
         logger.debug("status after event applied {}", aggregateRoot);
         events.add(event);
     }
 
+    public void emit(Command command) {
+        checkNotNull(command);
+        commands.add(command);
+        logger.debug("emitted command {} generated by an operation on {}", command, aggregateRoot);
+    }
+
     @Override
     public List<Event> getEvents() {
+        bus.unregister(aggregateRoot);
         return Collections.unmodifiableList(events);
     }
 
-    private void _applyEventsOn(AggregateRoot instance, Event event) {
-        try {
-            mm.invoke(instance, event);
-        } catch (Exception e) {
-            throw new RuntimeException("Error when executing with reflection", e.getCause());
-        }
+    public List<Command> getCommands() {
+        return Collections.unmodifiableList(commands);
     }
+
 }
