+++ myeslib2-jdbi/src/main/java/org/myeslib/jdbi/function/multimethod/MultiMethodInteractionContext.java
@@ -1,6 +1,5 @@
-package org.myeslib.jdbi.function;
+package org.myeslib.jdbi.function.multimethod;
 
-import com.google.common.eventbus.EventBus;
 import org.myeslib.core.AggregateRoot;
 import org.myeslib.core.Command;
 import org.myeslib.core.Event;
@@ -14,46 +13,41 @@ import java.util.List;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-public class StatefulEventBus implements InteractionContext {
+public class MultiMethodInteractionContext implements InteractionContext {
 
-    private static final Logger logger = LoggerFactory.getLogger(StatefulEventBus.class);
+    private static final Logger logger = LoggerFactory.getLogger(MultiMethodInteractionContext.class);
     private final AggregateRoot aggregateRoot;
-    private final EventBus bus;
+    private final MultiMethod mm ;
     private final List<Event> events;
     private final List<Command> commands;
 
-    public StatefulEventBus(AggregateRoot aggregateRoot) {
-        this.bus = new EventBus();
+    public MultiMethodInteractionContext(AggregateRoot aggregateRoot) {
+        this.mm = MultiMethod.getMultiMethod(aggregateRoot.getClass(), "on");
         this.events = new ArrayList<>();
         this.commands = new ArrayList<>();
         checkNotNull(aggregateRoot);
         this.aggregateRoot = aggregateRoot;
-        bus.register(aggregateRoot);
     }
 
     @Override
     public void apply(Event event) {
         checkNotNull(event);
         logger.debug("applying event {} generated by an operation on {}", event, aggregateRoot);
-        bus.post(event);
+        _applyEventsOn(aggregateRoot, event);
         logger.debug("status after event applied {}", aggregateRoot);
         events.add(event);
     }
 
-    public void emit(Command command) {
-        checkNotNull(command);
-        commands.add(command);
-        logger.debug("emitted command {} generated by an operation on {}", command, aggregateRoot);
-    }
-
     @Override
     public List<Event> getEvents() {
-        bus.unregister(aggregateRoot);
         return Collections.unmodifiableList(events);
     }
 
-    public List<Command> getCommands() {
-        return Collections.unmodifiableList(commands);
+    private void _applyEventsOn(AggregateRoot instance, Event event) {
+        try {
+            mm.invoke(instance, event);
+        } catch (Exception e) {
+            throw new RuntimeException("Error when executing with reflection", e.getCause());
+        }
     }
-
 }
